Made candidates viewMyApplication and recruiter can update JobApplication Status

+++ src/main/java/JobPortal/SpringJobPortal/Dto/CandidateProfileReqDto.java
@@ -30,7 +30,7 @@ public class CandidateProfileReqDto {
 
     @NotNull
     @Positive
-    private Long experience;
+    private Double experience;
 
     @NotBlank
     private String education;
